+++ src/objects/H2DE_bar_object.cpp
@@ -1,8 +1,8 @@
 #include "H2DE/H2DE_object.h"
 
 // INIT
-H2DE_BarObject::H2DE_BarObject(H2DE_Engine* engine, H2DE_ObjectData od, H2DE_BarObjectData b) : H2DE_Object(engine, od), bod(b), percentage(b.defaultPercentage) {
-
+H2DE_BarObject::H2DE_BarObject(H2DE_Engine* engine, H2DE_ObjectData od, H2DE_BarObjectData b) : H2DE_Object(engine, od), bod(b), value(b.defaultValue) {
+    refreshPercentage();
 }
 
 H2DE_BarObject* H2DE_CreateBarObject(H2DE_Engine* engine, const H2DE_ObjectData& od, const H2DE_BarObjectData& bod) {
@@ -28,8 +28,13 @@ void H2DE_BarObject::update() {
     }
 }
 
+void H2DE_BarObject::refreshPercentage() {
+    value = std::clamp(value, bod.min, bod.max);
+    percentage = (value - bod.min) / (bod.max - bod.min) * 100;
+}
+
 // GETTER
-std::vector<H2DE_SurfaceBuffer> H2DE_BarObject::getSurfaces() const {
+std::vector<H2DE_SurfaceBuffer> H2DE_BarObject::getSurfaceBuffers() const {
     H2DE_SurfaceBuffer backgroundBuffer = H2DE_SurfaceBuffer();
     backgroundBuffer.surface = bod.background;
     backgroundBuffer.offset = { 0.0f, 0.0f };
@@ -53,6 +58,28 @@ H2DE_Surface* H2DE_GetBarBackgroundSurface(const H2DE_BarObject* bar) {
 }
 
 // SETTER
-void H2DE_SetBarPercentage(H2DE_BarObject* bar, float percentage) {
-    bar->percentage = std::clamp(percentage, 0.0f, 100.0f);
+void H2DE_SetBarValue(H2DE_BarObject* bar, float value) {
+    bar->value = std::clamp(value, bar->bod.min, bar->bod.max);
+    bar->refreshPercentage();
+}
+
+void H2DE_SetBarValue(H2DE_BarObject* bar, float value, unsigned int duration, H2DE_Easing easing, bool pauseSensitive) {
+    value =  std::clamp(value, bar->bod.min, bar->bod.max);
+
+    const float defaultValue = bar->value;
+    const float valueToAdd = value - defaultValue;
+
+    H2DE_CreateTimeline(bar->engine, duration, easing, [bar, defaultValue, valueToAdd](float blend) {
+        H2DE_SetBarValue(bar, defaultValue + (valueToAdd * blend));
+    }, nullptr, 0, pauseSensitive);
+}
+
+void H2DE_SetBarMin(H2DE_BarObject* bar, float min) {
+    bar->bod.min = min;
+    bar->refreshPercentage();
+}
+
+void H2DE_SetBarMax(H2DE_BarObject* bar, float max) {
+    bar->bod.max = max;
+    bar->refreshPercentage();
 }