+++ src/objects/H2DE_basic_object.cpp
@@ -24,10 +24,14 @@ void H2DE_BasicObject::update() {
 }
 
 // GETTER
-std::vector<H2DE_SurfaceBuffer> H2DE_BasicObject::getSurfaces() const {
+std::vector<H2DE_SurfaceBuffer> H2DE_BasicObject::getSurfaceBuffers() const {
     H2DE_SurfaceBuffer buffer = H2DE_SurfaceBuffer();
     buffer.surface = bod.surface;
     buffer.offset = { 0.0f, 0.0f };
     buffer.size = od.size;
     return { buffer };
+}
+
+H2DE_Surface* H2DE_GetBasicObjectSurface(const H2DE_BasicObject* basicObject) {
+    return basicObject->bod.surface;
 }