+++ src/objects/H2DE_button_object.cpp
@@ -60,11 +60,43 @@ void H2DE_ButtonObject::update() {
     }
 }
 
+// CALLS
+void H2DE_HoverButton(const H2DE_ButtonObject* button) {
+    if (button->bod.onhover) {
+        button->bod.onhover();
+    }
+}
+
+void H2DE_OutButton(const H2DE_ButtonObject* button) {
+    if (button->bod.onout) {
+        button->bod.onout();
+    }
+}
+
+void H2DE_ClickButton(const H2DE_ButtonObject* button) {
+    if (button->bod.onclick) {
+        button->bod.onclick();
+    }
+}
+
 // GETTER
-std::vector<H2DE_SurfaceBuffer> H2DE_ButtonObject::getSurfaces() const {
+std::vector<H2DE_SurfaceBuffer> H2DE_ButtonObject::getSurfaceBuffers() const {
     H2DE_SurfaceBuffer buffer = H2DE_SurfaceBuffer();
     buffer.surface = bod.surface;
     buffer.offset = { 0.0f, 0.0f };
     buffer.size = od.size;
     return { buffer };
+}
+
+// SETTER
+void H2DE_SetButtonHover(H2DE_ButtonObject* button, const std::function<void()>& onhover) {
+    button->bod.onhover = onhover;
+}
+
+void H2DE_SetButtonOut(H2DE_ButtonObject* button, const std::function<void()>& onout) {
+    button->bod.onout = onout;
+}
+
+void H2DE_SetButtonClick(H2DE_ButtonObject* button, const std::function<void()>& onclick) {
+    button->bod.onclick = onclick;
 }